Refactor dashboard date & trend utils into lib
