+++ src/app/dashboard/(tabs)/people/page.tsx
@@ -1,6 +1,6 @@
-import { buildRangeFromPreset } from "@/components/dashboard/dashboard-filters";
 import { PeopleView } from "@/components/dashboard/people-view";
 import { getDashboardAnalytics } from "@/lib/dashboard/analytics";
+import { buildRangeFromPreset } from "@/lib/dashboard/date-range";
 import { fetchSyncStatus } from "@/lib/sync/service";
 
 export const dynamic = "force-dynamic";
