+++ src/lib/dashboard/metric-formatters.ts
